fix: fix compilation errors
- Add missing Canvas, Offset, Size imports
- Add missing remember, mutableIntStateOf imports
- Add AnimatedVisibility import
- Fix Long to Int type mismatch in applyNaturalDecay
- Fix fillMax -> fillMaxWidth
- Fix remember inside items() block
- Fix ActionButton repository creation

+++ android/app/src/main/java/com/timenw/cattracker/data/repository/CatRepository.kt
@@ -312,7 +312,7 @@ class CatRepository(private val context: Context) {
 
     fun applyNaturalDecay(cat: Cat): Cat {
         val hoursSinceLastInteraction = (System.currentTimeMillis() - cat.lastInteractionTime) / (1000 * 60 * 60)
-        val decayFactor = hoursSinceLastInteraction.coerceAtMost(24)
+        val decayFactor = hoursSinceLastInteraction.coerceAtMost(24).toInt()
         val newCat = cat.copy(
             hunger = (cat.hunger - decayFactor * 2).coerceIn(0, 100),
             happiness = (cat.happiness - decayFactor * 1).coerceIn(0, 100),
